+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 - Added back button in notification detailed view
+- Added alert detail view
 
 ### Changed
 - Changed Layout in notification detailed view